Fix import and move schema
