+++ doku/models/schemas/__init__.py
@@ -1,7 +1,7 @@
-from doku.models.resource import Resource
-from doku.models.schemas.document import DocumentSchema, VariableSchema
-from doku.models.schemas.variable import VariableGroupSchema
+from doku.models.schemas.document import DocumentSchema
+from doku.models.schemas.variable import VariableGroupSchema, VariableSchema
 from doku.models.schemas.template import TemplateSchema, StylesheetSchema
+from doku.models.schemas.resource import ResourceSchema
 from doku.models.schemas.snippet import SnippetSchema
 
 __all__ = [
@@ -10,6 +10,6 @@ __all__ = [
     VariableGroupSchema,
     TemplateSchema,
     StylesheetSchema,
-    Resource,
+    ResourceSchema,
     SnippetSchema
 ]
